[Common] code style

+++ src/tokenomics/contracts/base/FreezableToken.sol
@@ -1,18 +1,19 @@
+// SPDX-License-Identifier: GPL-2.0-or-later
 pragma solidity =0.7.6;
 
-import "../base/ERC20.sol";
+import '../base/ERC20.sol';
 
-abstract contract FreezableToken is ERC20{
+abstract contract FreezableToken is ERC20 {
     using SafeMath for uint256;
     // freezing chains
-    mapping (bytes32 => uint64) internal chains;
+    mapping(bytes32 => uint64) internal chains;
     // freezing amounts for each chain
-    mapping (bytes32 => uint) internal freezings;
+    mapping(bytes32 => uint256) internal freezings;
     // total freezing balance per address
-    mapping (address => uint) internal freezingBalance;
+    mapping(address => uint256) internal freezingBalance;
 
-    event Freezed(address indexed to, uint64 release, uint amount);
-    event Released(address indexed owner, uint amount);
+    event Freezed(address indexed to, uint64 release, uint256 amount);
+    event Released(address indexed owner, uint256 amount);
 
     /**
      * @dev Gets the balance of the specified address include freezing tokens.
@@ -41,11 +42,11 @@ abstract contract FreezableToken is ERC20{
      * @param _addr Address of freeze tokens owner.
      * @param _index Freezing portion index. It ordered by release date descending.
      */
-    function getFreezing(address _addr, uint _index) internal view returns (uint64 _release, uint _balance) {
-        for (uint i = 0; i < _index + 1; i++) {
+    function getFreezing(address _addr, uint256 _index) internal view returns (uint64 _release, uint256 _balance) {
+        for (uint256 i = 0; i < _index + 1; i++) {
             _release = chains[toKey(_addr, _release)];
             if (_release == 0) {
-                return (0,0);
+                return (0, 0);
             }
         }
         _balance = freezings[toKey(_addr, _release)];
@@ -93,7 +94,7 @@ abstract contract FreezableToken is ERC20{
         }
     }
 
-    function toKey(address _addr, uint _release) internal pure returns (bytes32 result) {
+    function toKey(address _addr, uint256 _release) internal pure returns (bytes32 result) {
         // increase entropy
         result = 0x5749534800000000000000000000000000000000000000000000000000000000;
         assembly {
@@ -114,7 +115,7 @@ abstract contract FreezableToken is ERC20{
         }
 
         bytes32 nextKey = toKey(_to, next);
-        uint parent;
+        uint256 parent;
 
         while (next != 0 && _until > next) {
             parent = next;
@@ -144,7 +145,11 @@ abstract contract FreezableToken is ERC20{
      * @param _until Release date, must be in future.
      * @return A boolean that indicates if the operation was successful.
      */
-    function mintAndFreeze(address _to, uint _amount, uint64 _until) internal returns (bool) {
+    function mintAndFreeze(
+        address _to,
+        uint256 _amount,
+        uint64 _until
+    ) internal returns (bool) {
         _totalSupply = _totalSupply.add(_amount);
 
         bytes32 currentKey = toKey(_to, _until);
